create library table
