+++ src/app/app.module.ts
@@ -73,6 +73,7 @@ import { SlrViewComponent } from './components/libraries/slr-view/slr-view.compo
 import { AddSlrDialogComponent } from './components/libraries/dialogs/add-slr-dialog/add-slr-dialog.component';
 import { MatTabsModule } from '@angular/material/tabs';
 import { SlrPropertiesComponent } from './components/libraries/slr-properties/slr-properties.component';
+import { LibraryTableComponent } from './components/libraries/library-table/library-table.component';
 
 @NgModule({
   declarations: [
@@ -101,6 +102,7 @@ import { SlrPropertiesComponent } from './components/libraries/slr-properties/sl
     SlrViewComponent,
     AddSlrDialogComponent,
     SlrPropertiesComponent,
+    LibraryTableComponent,
   ],
     imports: [
         MDBBootstrapModule.forRoot(),
