+++ src/app/components/libraries/library-table/library-table.component.html
@@ -1,0 +1,124 @@
+<div *ngIf="showTable">
+    <!-- Search -->
+    <div class="row">
+        <div class="col-12">
+            <div class="md-form m-2">
+                <input
+                        type="text"
+                        [(ngModel)]="searchText"
+                        (input)="onSearchChange()"
+                        class="form-control"
+                        id="search"
+                        mdbInput
+                />
+                <label for="search">Search</label>
+            </div>
+        </div>
+    </div>
+    <!-- Add-Button -->
+    <div class="row m-1">
+        <button mdbBtn class="control-button add-button btn-sm" type="button"
+                color="green"
+                outline="true"
+                (click)="onAddEntry()" mdbWavesEffect
+                [hidden]="!uiConfig.features.enableEditing">
+            <span class="material-icons">{{addIcon}}</span>
+        </button>
+        <!-- Submit-Selected-Button -->
+        <button mdbBtn *ngIf="selection.hasValue()" class="control-button delete-button btn-sm"
+                type="button" color=red outline="true"
+                (click)="onDeleteEntries()" mdbWavesEffect
+                [hidden]="!uiConfig.features.enableEditing">
+            <mat-icon>delete</mat-icon>
+        </button>
+        <!--            <button mdbBtn *ngIf="allowLink && selection.hasValue()" class="control-button link-button btn-sm"-->
+        <!--                    type="button" color=blue outline="true"-->
+        <!--                    (click)="onLinkSubmitted()" mdbWavesEffect-->
+        <!--                    [hidden]="!uiConfig.features.enableEditing">-->
+        <!--                <mat-icon>link</mat-icon>-->
+        <!--            </button>-->
+    </div>
+    <!-- Data-Table -->
+    <div class="table-responsive">
+        <table id="tableSortExample" mdbTable matSort (matSortChange)="sortData()">
+            <!-- Table-Header with the Header-Data -->
+            <thead class="data-header">
+            <tr>
+                <!-- Optional Select-All Checkbox Element in first Header-Column -->
+                <th class="align-middle center-text table-select">
+                    <div class="d-flex align-baseline justify-content-center">
+                        <mat-checkbox color="warn" (change)="$event ? masterToggle() : null"
+                                      [checked]="selection.hasValue() && isAllSelected()"
+                                      [indeterminate]="selection.hasValue() && !isAllSelected()"></mat-checkbox>
+                    </div>
+                </th>
+
+                <!-- One column for each needed data variable -->
+                <th *ngFor="let column of tableColumns; let i = index"
+                    class="th-lg hide-arrow align-middle table-data"
+                    [mdbTableSort]="entries"
+                    [sortBy]="variableNames[i]"
+                    mat-sort-header="{{ variableNames[i] }}"
+                    aria-controls="tableSortExample"
+                    scope="col">
+                    <a>{{ column }}</a>
+                </th>
+                <th class="align-middle table-actions"
+                    *ngIf="uiConfig.features.enableEditing">
+                    <a> Actions </a>
+                </th>
+            </tr>
+            </thead>
+            <!-- Table-Body with the Data-Elements -->
+            <tbody *ngIf="entries && entries.length > 0" align-middle justify-content-center>
+            <tr *ngFor="let dataEntry of entries"
+                [ngClass]="dataEntryIsDisabled(dataEntry) ? 'control-disabled' : 'data-element'">
+                <!-- Optional Select-Checkbox Element in first Body-Column -->
+                <th class="align-middle center-text table-select">
+                    <mat-checkbox color="warn"
+                                  [checked]="selection.isSelected(dataEntry)"
+                                  [disabled]="dataEntryIsDisabled(dataEntry)"
+                                  (click)="$event.stopPropagation()"
+                                  (change)="$event ? rowToggle(dataEntry) : null"></mat-checkbox>
+                </th>
+                <!-- One column for each needed data variable value -->
+                <td class="crop align-middle table-data" *ngFor="let variableName of variableNames"
+                    (click)="!dataEntryIsDisabled(dataEntry) && onElementClicked(dataEntry)">
+                    <!-- Apply URL-Class and Event if variable has a link -->
+                    <!--                                                <a [ngClass]="{'table-url': isLink(variableName)}"-->
+                    <!--                                                   (click)="isLink(variableName) && onUrlClicked($event, dataEntry, variableName)">-->
+                    <!--                                                    {{ dataEntry[variableName] }}-->
+                    <!--                                                </a>-->
+                    {{ dataEntry[variableName] }}
+                </td>
+                <td class="table-actions"
+                    *ngIf="!dataEntryIsDisabled(dataEntry) && uiConfig.features.enableEditing">
+                    <!--                            <button mat-icon-button [ngClass]="{'mr-1': true}" class="update-small"-->
+                    <!--                                    (click)="onUpdateClicked(dataEntry); $event.stopPropagation()">-->
+                    <!--                                <mat-icon>create</mat-icon>-->
+                    <!--                            </button>-->
+                    <button mat-icon-button [ngClass]="{'mr-1': true}"
+                            class="delete-small"
+                            (click)="onSingleDelete(dataEntry); $event.stopPropagation()">
+                        <mat-icon>delete</mat-icon>
+                    </button>
+                    <!--                        <button *ngIf="allowLink" mat-icon-button [ngClass]="{'mr-1': allowLink}" class="link-small"-->
+                    <!--                                (click)="onSingleLink(dataEntry); $event.stopPropagation()">-->
+                    <!--                            <mat-icon>link</mat-icon>-->
+                    <!--                        </button>-->
+                </td>
+                <!-- Empty table action if the entry is disabled -->
+                <td *ngIf="dataEntryIsDisabled(dataEntry)"></td>
+            </tr>
+            </tbody>
+            <!-- Empty Table-Body if no elements found -->
+            <tbody *ngIf="!entries || entries.length <= 0" align-middle justify-content-center>
+            <tr>
+                <td class="text-center" colspan="1000"><a>{{ emptyTableMessage }}</a></td>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+</div>
+
+
