+++ src/app/components/libraries/library-view/library-view.component.ts
@@ -1,4 +1,12 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import {
+  AfterViewInit,
+  Component,
+  EventEmitter,
+  Input,
+  OnInit,
+  Output,
+  ViewChild,
+} from '@angular/core';
 import { LibrariesService } from 'api-library/services/libraries.service';
 import { SelectionModel } from '@angular/cdk/collections';
 import { BibEntryDto } from 'api-library/models/bib-entry-dto';
@@ -15,28 +23,22 @@ import {
   ConfirmDialogComponent,
   ConfirmDialogData,
 } from '../../generics/dialogs/confirm-dialog.component';
+import { LibraryTableComponent } from '../library-table/library-table.component';
 
 @Component({
   selector: 'app-library-view',
   templateUrl: './library-view.component.html',
   styleUrls: ['./library-view.component.scss'],
 })
-export class LibraryViewComponent implements OnInit {
+export class LibraryViewComponent implements OnInit, AfterViewInit {
   @Input() addIcon = 'playlist_add';
   @Input() emptyTableMessage = 'No elements found';
-  @Output() elementClicked = new EventEmitter<any>();
-  @Output() updateClicked = new EventEmitter<any>();
+  @ViewChild(LibraryTableComponent)
+  private libraryTable!: LibraryTableComponent;
   selection = new SelectionModel<any>(true, []);
-  entries: TableEntry[] = [];
-  allEntries: TableEntry[] = [];
-  searchText = '';
-  tableColumns = ['Cite Key', 'Title', 'Authors', 'Entry Type', 'Date'];
-  variableNames = ['id', 'title', 'author', 'entrytype', 'date'];
   libraries$: string[];
   library: string;
-  disabledDataEntries: Set<any> = new Set<any>();
   librariesExist = false;
-
   uiConfig: QcAtlasUiConfiguration;
 
   constructor(
@@ -47,6 +49,9 @@ export class LibraryViewComponent implements OnInit {
 
   ngOnInit(): void {
     this.uiConfig = this.configService.configuration;
+  }
+
+  ngAfterViewInit(): void {
     this.getLibraryNames();
   }
 
@@ -58,32 +63,12 @@ export class LibraryViewComponent implements OnInit {
           lib.substr(0, lib.length - 4)
         );
         this.library = this.libraries$[0];
-        this.getLibrary(this.library);
+        this.libraryTable.getLibrary(this.library);
         this.librariesExist = true;
       } else {
         this.librariesExist = false;
       }
     });
-  }
-
-  getLibrary(libraryName: string): void {
-    this.entries = [];
-    this.allEntries = [];
-    this.library = libraryName;
-    this.libraryService
-      .getLibraryEntries({ libraryName: this.library })
-      .subscribe((bibentries) => {
-        bibentries.bibEntries.forEach((entry) => {
-          this.entries.push({
-            id: entry.citationKey,
-            author: entry.author,
-            title: entry.title,
-            entrytype: entry.entryType,
-            date: entry.date,
-          });
-        });
-        this.allEntries = this.entries;
-      });
   }
 
   onAddLibrary(): void {
@@ -103,7 +88,7 @@ export class LibraryViewComponent implements OnInit {
               );
               this.library = dialogResult.name;
               this.librariesExist = true;
-              this.getLibrary(this.library);
+              this.libraryTable.getLibrary(this.library);
             });
             this.utilService.callSnackBar(
               'Successfully added the library "' + this.library + '".'
@@ -115,53 +100,6 @@ export class LibraryViewComponent implements OnInit {
             );
           }
         );
-      });
-  }
-
-  onElementClicked(element): void {
-    this.elementClicked.emit(element);
-    this.selection.clear();
-    this.libraryService
-      .getBibEntryMatchingCiteKey({
-        citeKey: element.id,
-        libraryName: this.library,
-      })
-      .subscribe((bibEntry) => {
-        this.utilService
-          .createDialog(AddBibentryDialogComponent, {
-            title: 'Update bibTeX entry',
-            bibEntry,
-          })
-          .afterClosed()
-          .subscribe((dialogResult) => {
-            if (dialogResult.bibEntry) {
-              const bibEntryDto = dialogResult.bibEntry as BibEntryDto;
-              this.libraryService
-                .updateEntry({
-                  citeKey: bibEntryDto.citationKey,
-                  libraryName: this.library,
-                  body: bibEntryDto,
-                })
-                .subscribe(() => this.getLibrary(this.library));
-            }
-          });
-      });
-  }
-
-  onAddEntry(): void {
-    this.utilService
-      .createDialog(AddBibentryDialogComponent, {
-        title: 'Add a new bibTeX entry',
-        bibEntry: {},
-      })
-      .afterClosed()
-      .subscribe((dialogResult) => {
-        if (dialogResult.bibEntry) {
-          const bibEntryDto = dialogResult.bibEntry as BibEntryDto;
-          this.libraryService
-            .addEntryToLibrary({ libraryName: this.library, body: bibEntryDto })
-            .subscribe(() => this.getLibrary(this.library));
-        }
       });
   }
 
@@ -196,157 +134,9 @@ export class LibraryViewComponent implements OnInit {
       });
   }
 
-  onDeleteEntries(): void {
-    const dialogData: ConfirmDialogData = {
-      title: 'Confirm Deletion',
-      message: 'Are you sure you want to delete the following entries:',
-      data: this.selection.selected,
-      variableName: 'id',
-      yesButtonText: 'yes',
-      noButtonText: 'no',
-    };
-    this.utilService
-      .createDialog(ConfirmDialogComponent, dialogData)
-      .afterClosed()
-      .subscribe((dialogResult) => {
-        if (dialogResult) {
-          const deletionTasks = [];
-          const snackbarMessages = [];
-          let successfulDeletions = 0;
-          this.selection.selected.forEach((element) => {
-            deletionTasks.push(
-              this.libraryService
-                .deleteEntryFromLibrary({
-                  citeKey: element.id,
-                  libraryName: this.library,
-                })
-                .toPromise()
-                .then(() => {
-                  successfulDeletions++;
-                  snackbarMessages.push(
-                    'Successfully deleted entry "' + element.id + '".'
-                  );
-                })
-                .catch(() => {
-                  snackbarMessages.push(
-                    'Could not delete entry "' + element.id + '".'
-                  );
-                })
-            );
-          });
-          forkJoin(deletionTasks).subscribe(() => {
-            this.getLibrary(this.library);
-            this.selection.clear();
-            this.utilService.callSnackBar(
-              'Successfully deleted ' +
-                successfulDeletions +
-                '/' +
-                dialogResult.data.length +
-                ' entries.'
-            );
-            this.utilService.callSnackBarSequence(snackbarMessages);
-          });
-        }
-      });
+  changeLibrary(library: string): void {
+    this.libraryTable.getLibrary(library);
+    this.library = library;
   }
 
-  sortData(): void {
-    this.selection.clear();
-  }
-
-  isAllSelected(): boolean {
-    return (
-      this.entries.length ===
-      this.selection.selected.length + this.disabledDataEntries.size
-    );
-  }
-
-  // Toggle all check boxes
-  masterToggle(): void {
-    const isAllSelected = this.isAllSelected();
-    this.entries.forEach((element) => {
-      if (!this.dataEntryIsDisabled(element)) {
-        this.changeSelection(element, !isAllSelected);
-      }
-    });
-  }
-
-  dataEntryIsDisabled(dataEntry: any): boolean {
-    return this.disabledDataEntries.has(dataEntry.id);
-  }
-
-  rowToggle(row: any): void {
-    this.changeSelection(row, !this.selection.isSelected(row));
-  }
-
-  changeSelection(row: any, select: boolean): void {
-    if (select !== this.selection.isSelected(row)) {
-      this.selection.toggle(row);
-    }
-  }
-
-  onUpdateClicked(element): void {
-    this.updateClicked.emit(element);
-    this.selection.clear();
-  }
-
-  onSingleDelete(element): void {
-    const dialogData: ConfirmDialogData = {
-      title: 'Confirm Deletion',
-      message: 'Are you sure you want to delete the following entry:',
-      data: [element],
-      variableName: 'id',
-      yesButtonText: 'yes',
-      noButtonText: 'no',
-    };
-    this.utilService
-      .createDialog(ConfirmDialogComponent, dialogData)
-      .afterClosed()
-      .subscribe((dialogResult) => {
-        if (dialogResult) {
-          this.libraryService
-            .deleteEntryFromLibrary({
-              citeKey: element.id,
-              libraryName: this.library,
-            })
-            .toPromise()
-            .then(() => {
-              this.utilService.callSnackBar(
-                'Successfully deleted entry "' + element.id + '".'
-              );
-              this.getLibrary(this.library);
-            })
-            .catch(() => {
-              this.utilService.callSnackBar(
-                'Could not delete entry "' + element.id + '".'
-              );
-            });
-        }
-      });
-  }
-
-  onSearchChange(): void {
-    this.entries = this.allEntries;
-    this.entries = this.entries.filter((entry) => {
-      const term = this.searchText.toLowerCase();
-      return (
-        entry.entrytype.toLowerCase().includes(term) ||
-        entry.id.toLowerCase().includes(term) ||
-        entry.author?.toLowerCase().includes(term) ||
-        entry.keywords?.toLowerCase().includes(term) ||
-        entry.date?.toLowerCase().includes(term) ||
-        entry.title?.toLowerCase().includes(term)
-      );
-    });
-    this.selection.clear();
-  }
-}
-
-interface TableEntry {
-  id: string;
-  author?: string;
-  title?: string;
-  entrytype?: string;
-  date?: string;
-  keywords?: string;
 }
